Remove ome- metdata handling from zarr
