+++ modules/zarr/src/index.ts
@@ -1,3 +1,1 @@
-export { loadBioformatsZarr, loadOmeZarr } from './zarr';
-
-export { default as ZarrPixelSource } from './zarr/pixel-source';
+export { load } from './lib';
