+++ modules/zarr/test/index.js
@@ -1,4 +1,2 @@
 import './utils.spec';
-
-import './bioformats-zarr.spec';
-import './zarr-lib.spec';
+import './pixel-source.spec';
